+++ crates/forge/tests/e2e/mod.rs
@@ -4,3 +4,4 @@ mod collection;
 mod file_operations;
 mod fuzzing;
 mod running;
+mod workspaces;